feat(cardstats): use card components and move
