+++ src/pages/Dashboard.tsx
@@ -1,8 +1,4 @@
-import CardLineChart from 'components/Cards/CardLineChart';
-import CardBarChart from 'components/Cards/CardBarChart';
-import CardPageVisits from 'components/Cards/CardPageVisits';
-import CardSocialTraffic from 'components/Cards/CardSocialTraffic';
-import CardStats from 'components/Cards/CardStats';
+import CardStats from 'components/CardStats';
 
 const Dashboard: React.FC = () => {
   return (
